It is now game over when 5 flies are eaten
Also added a score count and made the game end after the animation of the tongue coming back. reworked a bit on the gameover screen and made it so that you can restart the game by pressing space bar. Now the only thing left is to make the lily pad untouchable