added a second floor
added a second floor and made it so that if it overlaps with the hero (cube) well the hero stays on the top of the floor. Also added a function to have a smooth animation of the hero falling off the platform. Now I need to put the floor as an array, so that it can do the same calculations with each floor I will add :)